# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для управления автомобилями",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Список автомобилей пользователя",
                "operationId": "4e913ee0b6125f5ad41f32df1df5b8a4",
                "responses": {
                    "200": {
                        "description": "Список автомобилей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CarModel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Создать автомобиль",
                "operationId": "fbbced9da0be8a33be9dfad182f803bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_model_id",
                                    "brand_id",
                                    "year",
                                    "color",
                                    "mileage"
                                ],
                                "properties": {
                                    "car_model_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2020
                                    },
                                    "mileage": {
                                        "type": "integer",
                                        "example": 15000
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "red"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Автомобиль создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cars/{car}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Показать автомобиль",
                "operationId": "91fb8ee0e3137745bcabf3a8bd78c386",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные автомобиля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Удалить автомобиль",
                "operationId": "0fc1699e0e5ba207ce31b7e1dda1ba9f",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Автомобиль удален"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Cars"
                ],
                "summary": "Обновить автомобиль",
                "operationId": "0bd03358c1fc2590b669e2f077888246",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_model_id",
                                    "brand_id",
                                    "year",
                                    "color",
                                    "mileage"
                                ],
                                "properties": {
                                    "car_model_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2020
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "red"
                                    },
                                    "mileage": {
                                        "type": "integer",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Автомобиль обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarModel": {
                "title": "Car",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "year": {
                        "type": "integer",
                        "example": 2025
                    },
                    "mileage": {
                        "type": "integer",
                        "example": 20000
                    },
                    "color": {
                        "type": "string",
                        "example": "black"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}